requirements.txt
elasticsearch
sentence-transformers
Pillow
matplotlib
IPython
fastapi
python-multipart
deepface